app/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
astronomer-cosmos==1.6.0
apache-airflow-providers-postgres
apache-airflow-providers-airbyte
apache-airflow-providers-airbyte[http]
google-cloud-bigquery-storage